snapshot_share/requirements.txt
aws-cdk-lib==2.43.1
aws-cdk.aws-lambda-python-alpha==2.43.1a0
constructs>=10.0.0,<11.0.0
